refactor(core): move persisted metric contract types to @lobu/core
The server compiles/validates stored metric JSON and must not import @lobu/cli,
so the persisted shapes belong in core (pi review). New packages/core/src/metrics.ts
holds EventSet/Measure/Dimension/Segment/MetricReadMode/MetricTier/FactMatchRule +
the federation shapes EntityTypeContribution/ReflectedMeasure. cli/config and
connector-sdk now re-export from core for authoring. Plain interfaces; the zod
validator (core's lib, not Typebox) lands with the persistence path that uses it.
