+++ packages/cli/src/config/define.ts
@@ -10,6 +10,7 @@
  * config and references.
  */
 
+import type { Dimension, EventSet, Measure, Segment } from "@lobu/core";
 import type {
   ConnectorClass,
   ConnectorRuntime,
@@ -51,108 +52,19 @@ export interface EntityBacking {
 }
 
 // ---------------------------------------------------------------------------
-// Entity-bound metrics (declarative; the compiler lowers these to backing SQL)
+// Entity-bound metrics — the contract types live in `@lobu/core` (shared by CLI
+// authoring, connector federation, and server compile/validate). Re-exported
+// here so configs can import them alongside `defineEntityType`.
 // ---------------------------------------------------------------------------
-
-/**
- * Temporal truth a measure reads over the append-only event stream:
- * - `"current"` — `current_event_records` (superseded rows masked); the default.
- * - `"raw"`     — `events` verbatim, including superseded rows (audit/debug).
- * - `{ asOf }`  — point-in-time snapshot; `asOf` is an ISO-8601 instant compared
- *                 against `occurred_at` (event time, NOT system time). A richer
- *                 structured form (relative durations, system-time) is deferred.
- */
-export type MetricReadMode = "current" | "raw" | { asOf: string };
-
-/**
- * Cross-source fact-identity rule. DEFERRED — a no-op until a measure fuses more
- * than one source. Shaped now so the {@link EventSet} schema needs no breaking
- * change when the first multi-source measure ships; until then use
- * {@link EventSet.dedupeKey} for same-source dedupe.
- *
- * Domain-agnostic: keys are logical field names, not a fixed finance tuple.
- */
-export interface FactMatchRule {
-  /**
-   * Per-source field map normalizing heterogeneous rows to a common fact tuple:
-   * logical field name → SQL expression. e.g.
-   * `{ amount: "metadata->>'amount'", at: "metadata->>'date'" }`.
-   */
-  key: Record<string, string>;
-  /** Per-key match tolerance, e.g. `{ at: "2d", amount: "0" }` (string; the compiler interprets). */
-  tolerance?: Record<string, string>;
-  /** Source priority when the same fact is seen twice (e.g. `["revolut","gmail"]`). */
-  prefer?: string[];
-}
-
-/**
- * A NAMED event set — how events resolve to this entity, at an explicit grain
- * (the join key). An entity can resolve events in several roles/grains, so the
- * grain is named, not implicit. The compiler lowers this to the base relation a
- * measure aggregates over (resolve → reads-mask → dedupe → segment).
- *
- * v1 implements `by: "alias"` only; `"window"`/`"link"` and `factIdentity` are
- * deferred. `where` predicates are raw SQL — parsed/validated/org-scoped by the
- * compiler, never trusted verbatim.
- */
-export interface EventSet {
-  /** alias: `field` ∈ entity aliases · window: `occurred_at` ∈ [start,end] · link: `events[].entity_ids`. */
-  by: "alias" | "window" | "link";
-  /** by:"alias" — event field matched against the entity's alias array. */
-  field?: string;
-  against?: "aliases";
-  /** by:"window" — entity property names bounding the time window. */
-  start?: string;
-  end?: string;
-  /** by:"window" — forbid an event attaching to more than one entity. */
-  cardinality?: "one_per_event";
-  /** Raw SQL predicate scoping the grain. */
-  where?: string;
-  /** Temporal truth to read. Default `"current"` (current_event_records). */
-  reads?: MetricReadMode;
-  /** Same-source identity: DISTINCT over this SQL-expression tuple, applied before aggregate. */
-  dedupeKey?: string[];
-  /** Cross-source identity — deferred (see {@link FactMatchRule}). */
-  factIdentity?: FactMatchRule;
-}
-
-export type MetricTier = "gold" | "silver" | "bronze";
-
-/** A governed aggregation bound to a named {@link EventSet} grain. */
-export interface Measure {
-  /** Name of the {@link EventSet} (on this entity) this measure aggregates over. */
-  eventSet: string;
-  agg: "sum" | "count" | "min" | "max" | "count_distinct";
-  /** SQL expression to aggregate. Required for every `agg` except `count` (validated at apply). */
-  expr?: string;
-  /** Extra raw-SQL predicate applied to this measure only. */
-  where?: string;
-  /** Names of {@link Segment}s (on this entity) to AND in, each applied per its `appliedBefore`. */
-  segments?: string[];
-  /** Dimensions safe to group WITH this measure; omitted ⇒ unknown (the compiler treats conservatively). */
-  safeDimensions?: string[];
-  /** REQUIRED — powers keyword discovery in the metric catalog. */
-  description: string;
-  owner?: string;
-  tier?: MetricTier;
-}
-
-/** A governed group-by. */
-export interface Dimension {
-  expr: string;
-  description: string;
-}
-
-/** A reusable named population filter (Anthropic "segment"). */
-export interface Segment {
-  description: string;
-  /** Raw SQL predicate. */
-  where: string;
-  /** Grain the filter applies at. */
-  on: "event" | "entity";
-  /** Ordering relative to dedupe. Default `"aggregate"`. */
-  appliedBefore?: "dedupe" | "aggregate";
-}
+export type {
+  Dimension,
+  EventSet,
+  FactMatchRule,
+  Measure,
+  MetricReadMode,
+  MetricTier,
+  Segment,
+} from "@lobu/core";
 
 export interface EntityType {
   readonly kind: "entityType";
